requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.6
bootstrap4==0.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-braces==1.13.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
Faker==1.0.7
idna==2.8
misaka==2.1.1
oauthlib==3.0.2
Pillow==6.0.0
pycparser==2.19
Pygments==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.25.3
wincertstore==0.2
